Update unit test

+++ mmengine/visualization/utils.py
@@ -177,7 +177,7 @@ def wait_continue(figure, timeout: float = 0, continue_key: str = ' ') -> int:
     https://github.com/matplotlib/matplotlib/blob/v3.5.x/lib/matplotlib/_blocking_input.py
 
     Args:
-        timeout (int): If positive, continue after ``timeout`` seconds.
+        timeout (float): If positive, continue after ``timeout`` seconds.
             Defaults to 0.
         continue_key (str): The key for users to continue. Defaults to
             the space key.
